requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
azure-functions==1.12.0
Babel==2.11.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.2
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
h11==0.14.0
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.2
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.4.8
jupyter-console==6.4.4
jupyter-core==5.1.1
jupyter-events==0.5.0
jupyter-server==2.0.5
jupyter-server-terminals==0.4.3
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.17.0
jupyterlab-widgets==3.0.5
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
mypy==1.0.0
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook==6.5.2
notebook-shim==0.2.2
numpy==1.19.2
openpyxl==3.0.10
outcome==1.2.0
packaging==22.0
pandas==1.1.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
platformdirs==2.6.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.8.0
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.6
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
selenium==4.7.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.8.0
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
virtualenv==20.17.1
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.5
wsproto==1.2.0
xlrd==2.0.1
zipp==3.11.0
